requirements.txt
browser-use==0.5.11
pyperclip==1.9.0
json-repair==0.44.1
MainContentExtractor==0.0.4
langchain-community
fastapi==0.103.0
langchain-openai==0.3.11
uvicorn ==0.35.0
python-dotenv == 1.1.1
uv ==0.8.11
asyncpg == 0.30.0
sqlalchemy == 2.0.31
psycopg2-binary == 2.9.10
pydantic == 2.11.7